pitch, yaw should be byte (r1.4 -> r1.3)

+++ src/main/java/com/github/dirtpowered/dirtmv/network/versions/Release51To39/ProtocolRelease51To39.java
@@ -224,8 +224,8 @@ public class ProtocolRelease51To39 extends ServerProtocol {
                         data.read(2),
                         data.read(3),
                         data.read(4),
-                        set(Type.BYTE, 0), // yaw
-                        set(Type.BYTE, 0), // pitch
+                        set(Type.BYTE, (byte) 0), // yaw
+                        set(Type.BYTE, (byte) 0), // pitch
                         data.read(5)
                 });
             }
